Adapt recent PCL package reorganization
