--- test/data/README.md
@@ -1,3 +1,0 @@
-# Test data
-
-Some pcd files from https://github.com/PointCloudLibrary/data for testing.
